Bugfix: Incorrect mapping
the anime-list mapping file had a tvdb id as valid id (integer) and an
imdb id. with a valid tvdb id the generator considered it to be a
TV-Show.

I added a check if the item has a imdb id to lookup the tmdb id with the
imdb id instead.