Updated generator
the generator will now merge the lists on its own. The reason for this
is that we need to have the Type to do requests to TheMovieDB if, for
example, tmdb Id is missing. Since both Series and Movies could share
the same ID this is now limited to elements with the "MOVIE" type only.