Fix #6 missing JSON response elements break generator
When there is a 404 error happening while the generator tries to get a TMDB ID with an IMDB ID, there is a JSONException because it cannot find the movie_results or tv_results of the response. The generator now checks if those elements are actually available before continuing to prevent the generator from running into the exception and failing the whole process.