Fix #7 added new sources and prevent null_id as key name
newly added sources in the AnimeOfflineDatabase would not be recognized and produced a 'null_id' key name that would then be added to the anime item which is useless especially since any unrecognized website would write and overwrite that element.

Two fixes were applied:
First, the new sources were added to the KeyNameUtils so that they get a correct key name.
The Second is that the KeyNameUtils.getValueWithId now either returns the correct key name or, if it isn't found, NULL. This should then prevent the null_id from being added in the future.